Add files via upload
This is a very basic project of applying the classification models on Historical Sales and active inventory dataset.

The file contains historical sales data (identified with the column titled File_Type). The historical data shows sales for the past 6 months. The binary target (1 = sale, 0 = no sale in past six months) is likely the primary target that should drive the analysis

•	SKU_number: This is the unique identifier for each product.
•	Order: Just a sequential counter. Can be ignored.
•	SoldFlag: 1 = sold in past 6 mos. 0 = Not sold
•	MarketingType = Two categories of how we market the product. This should probably be ignored, or better yet, each type should be considered independently.
•	NewReleaseFlag = Any product that has had a future release (i.e., Release Number > 1)